requirements.txt
cachetools==5.2.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
distlib==0.3.4
filelock==3.6.0
Flask==2.1.2
Flask-MySQLdb==1.0.1
google-auth==2.6.6
google-auth-oauthlib==0.5.1
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysqlclient==2.1.0
oauthlib==3.2.0
platformdirs==2.5.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
urllib3==1.26.9
virtualenv==20.14.1
Werkzeug==2.1.2
